Piccoli cambiamenti

+++ Distributore_Bevande/src/distributore/Macchinetta.java
@@ -4,13 +4,13 @@ import java.util.ArrayList;
 
 public class Macchinetta
 {
-	public int resto;
+	public double resto;
 	public int zucchero;
 	public int bacchette;
 	public int bicchieri;
 	public ArrayList<Prodotto> prodotti;
 
-	public Macchinetta(ArrayList<Prodotto> prodotti, int resto, int zucchero, int bacchette, int bicchieri) //Costruttore macchinetta
+	public Macchinetta(ArrayList<Prodotto> prodotti, double resto, int zucchero, int bacchette, int bicchieri) //Costruttore macchinetta
 	{
 		this.resto = resto;
 		this.zucchero = zucchero;
